GameStarterInterface, Starter{Proton,Wine}: Add gamestarter.py and move there
